requirements.txt
pandas
numpy
scikit-learn
seaborn
openai
google-genai
ollama
langchain
langchain-community
langchain-openai